removes alter table logic
Originally tried to implement the alter table algorithm presented in the
SQLite docs at:

https://www.sqlite.org/lang_altertable.html

This algorithm allowed the modification and removal of columns in a table
which SQLite does not provide syntax for.  This method worked; however, it
required that the foreign key constraints -- which are set per connection --
be turned off before the table is modified and turned back on when the
modification is finished.  The problem is that Ecto starts a transaction
before the alter statement is issued which prevents the foreign key
constraints from being changed.  This can cause an error when the table is
modified during the course of the algorithm.

In order to get around this problem I will probably need to implement foreign
key constraints with triggers as outlined here:

http://justatheory.com/computers/databases/sqlite/foreign_key_triggers.html

The above might be a difficult implementation so instead of holding up the
release of 0.1.0, I will just raise ArgumentErrors if the user tries to modify
or remove columns.  Once 0.1.0 is released, I will file an enhancement to put
the alter table functionality back in.